FR(endpoint): Search any user by username. Create a post
'SearchEndPoint' requires in the url the username, e.g search/myUsername

It returns a JSON with all the data of the required user.
Path is probably subject to change.

'CreatePostEndpoint' uses the path posts/create.
it needs an **UserID** that in the future will be retrieved automatically using the JWT, **parentID** this is necessary if the post is a reply or a comment to another post, **quoteID** to be used when the user wants to quote another user's post, and finally **body** the actual content of the post being created.

Ref Issue: #26